requirements.txt
# LangChain 0.3.x - Latest stable versions
langchain>=0.3.25
langchain-openai>=0.3.18
langchain-community>=0.3.24
langchain-core>=0.3.61
langchain-text-splitters>=0.3.8

# Core dependencies
openai>=1.82.0
pydantic>=2.11.5
pydantic-settings>=2.9.1
python-dotenv>=1.0.0
tiktoken>=0.9.0

# Web interface (Streamlit)
streamlit>=1.29.0

# Optional: For enhanced functionality
httpx>=0.28.0
httpx-sse>=0.4.0 